Refactored Code
Pulled out CannonBase functionality to it's own file
Pulled out CannonBarrel functionality to it's own file
Pulled out Tower functionality to it's own file
Started building projectile functionality
